server side async

+++ news-feed-server/src/main/java/com/learning/news_feed/dao/NewsRepository.java
@@ -3,5 +3,5 @@ package com.learning.news_feed.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.learning.news_feed.entity.News;
 
-public interface NewsRepository extends JpaRepository<News, Integer>{
+public interface NewsRepository extends JpaRepository<News, Long>{
 }
